requirements.txt
Flask==1.0.2
flask-wtf==0.14.3
Flask-WTF==0.14.3
flask-sqlalchemy==2.4.0
psycopg2==2.8.3
ipython==6.2.1
flask-bcrypt==0.7.1
flask-login==0.4.1
flask_script==2.0.6
flask_migrate==2.5.3
email_validator==1.1
Pillow==7.1.0
flask-mail==0.9.1

